Client: Attempted to find out why SyntaxTree.loadExpression fails for some tests
This branch is set up to almost exactly mirror the code in the `QueryUpdate` and the `QuerySelect` loading paths. It even leaves out the intermediate loading step using `QueryAssign` for `QueryUpdate` by copying it verbatim. And still all tests regarding `QuerySelect` work just fine but all tests regarding `QueryUpdate` fail with `TypeError: base_1.SyntaxTree.loadExpression is not a function`.