Server and Client: CodeResources may now be saved but ...
... the current solution is a not-so-nice hack. To bridge the gap
between the "snakeCase" requests to a server that prefers to get
"camelCase" things an *explicit* conversion is used. This is
because (in theory) it *may* be possible that there are legit use-cases
to have snake_cased names in the JSON payloads. The currently existing
models (BlockLanguage & SyntaxTree) at least do not make *any* attempt
to forbid keys with underscores.