Client: Refactored `CachedRequest<T>` to be triggered by a `BehaviorSubject`
Many thanks to Timm Hoffmann who came up with that idea.

The only sort of remaining problem is the `tap` side-effects for the
counter would trigger the Angular change detection. Maybe there is a
smarter way to work around this then the book I wrote in the comments
to describe the workaround.