Client, Server, Data: User provided generator instructions (WIP)
The `BlockLanguageDocument` now allows storage of a single
`BlockLanguageGeneratorDocument` which is then used for the actual
generation process. From now on the previously available static
generation options in `generator.description.ts` have been moved into
the database.

The instructions do have a substantial flaw though: Multi-blocks
implicitly rely on the ordering of keys `JSON`-objects. This will
require some reworking in the next commit but for the moment this
commit is massive (and breaking) enough as it is.