Client: Grammars may now specify "between" terminals
For "real world grammars" this approach is, to say the least,
unusual. Usually grammars would specify rules that explicitly model
out the separation character:

  list ::= item | item ',' list

But the motivation behind our grammars is slightly different: They
need to define tree structures (which have no concept of terminal
symbols) and are used as an input for block language generation (which
require exactly the "what goes between these items"-information). So I
have the hope that this hack will work nice enough.