Client: ParameterReferenceable now applies recursively
This means that all nested properties of `Instructions` are (formally)
able to be references. My gut says this is a far too wide-ranging
approach but for the moment (until proper specification and
documentation is available) this will stay in place.

Sadly the schema generator is not yet able to handle conditional
types, the current schemas have therefore been generated with a
patched generator that simply emits `any` for each conditional
type. My feeble attempt to hack in support can be seen at [0].

[0] https://github.com/vega/ts-json-schema-generator/pull/39