Client: Trucklino programs now inform about execution progress
The chosen approach "litters" the generated code with progress
callbacks. This doesn't make the code any nicer to read, but for the
moment this is what we have.

Known issue: The displayed blocks are only updated when browser events
are fired (e.g. the mouse is moved). This is probably due to the
explicit change detection.