Client: Broken mess with attempted introduction of ResourceReferencesService
The grammar gallery wants to inject the block language that must be
used for rendering via the components @Input, but some parts of the
rendering process still rely on the BlockLanguage that is part of the
CodeResource. This is especially a problem during validation, as the
validator is constructed based on the programming language in the
block language.

The comitted state is in the transition of passing the BlockLanguage
down the whole component chain, but this will result in a multitude of
problems with validation: The elements should share the same
validation results, which means they must be somewhat global ...